Implement compared value extraction.
Extracting the values to compare is a separate concern from comparing
them. I'm constantly fussing with the shape of comparator arguments in
the b-tree and MVCC libraries. Ascension binds a single extraction
strategy but I'm often overriding that strategy, so now Ascension will
bind no extraction strategy and we'll whittle arguments as necessary.

Closes #6.